Stop floating image appearing behind CTA
On tablets, news images are floated right, and text is expected to wrap
around them. However the “Call to action” block has a grey background
and covers the full width, this obscures the image and looks broken.

This only happens if the CTA is near the top of the article.
Example:
https://www.gov.uk/government/news/new-public-sector-boost-for-uk-steel