Delete redundant WLNA 'en' edition_translations
WLNA have an option to set an edition to be "designated as a foreign
language only". This has the effect of creating an edition with
`primary_locale` equal to something other than `en`. The
edition_translation table is then populated with a single row for this
locale. The slug for the document is set to the document_id and the
`base_path` then becomes document_id.locale (e.g.
/government/world-location-news/171142.es-419).

For some reason about 690 WLNA with `primary_locale` other than `en` had
a spurious row in the edition_translations table with the locale set to
`en`. For the most part the `title` and `description` were nil or empty
strings and the body was set to the same text as the primary_locale
text.

These spurious rows were causing sync check failures.

This migration deletes these rows.