Only run Sidekiq jobs under test when needed
Makes it easier to test by not having to deal with side effects of
background processing.

Using the default Sidekiq fake testing mode pushed jobs onto the queue
but does not execute them. We can then use the inline testing mode for
test cases where we want the work on the queue to complete before
asserting expected behaviour.